Added project notebook
Initial code upload. Current build is able to work through a folder of individual frames and extract differences, then compile the frame differences into a single output file equal to the size of a single frame. Frame difference blocks are lined with a two pixel border stretching the block colors to the edges. This is to reduce the influence that each block takes from their neighboring blocks during the upscaling process.